refactor(heatmap): extract window resolution to window.rs (#216)
Move preset_to_since_secs, resolve_effective_config, build_window_info,
format_epoch, and days_to_ymd from mod.rs to a dedicated window.rs module.
Also relocate the 14 window tests and 2 helpers (base_window, NOW).

All functions are pub(super). format_epoch and resolve_effective_config are
re-exported into mod.rs via use window::{...}. preset_to_since_secs is
internal to window.rs (only called by resolve_effective_config).
